corrected path to deleted bug

+++ helper/deleteUploadedFiles.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 
 module.exports = (filePath) => {
-  if (!filePath) return
+  if (filePath || filePath === null) return
   fs.unlinkSync(filePath);
   return true
 };
